+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to Knot.x will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
+ - [PR-427[(https://github.com/Cognifide/knotx/pull/427) - HttpRepositoryConnectorProxyImpl logging improvements
  - [PR-422](https://github.com/Cognifide/knotx/pull/422) - Configurable Handlebars delimiters
 
 ## 1.3.0